requirements.txt
appnope==0.1.2
asgiref==3.3.1
backcall==0.2.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.5
django-crispy-forms==1.10.0
django-extensions==3.1.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
gunicorn==20.0.4
ipython==7.10.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.10
psycopg2-binary==2.8.6
ptyprocess==0.7.0
Pygments==2.7.3
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
